feat(frontend): working tailwind
