+++ frontend/src/app/layout.tsx
@@ -1,23 +1,20 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
 import "./globals.css";
+import Head from "next/head";
+import { ReactNode } from "react";
 
 
-const inter = Inter({ subsets: ["latin"] });
+type LayoutProps = {
+  children?: ReactNode
+}
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+export const Layout = ({ children}: LayoutProps) => {
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
+  return <div>
+    <Head>
+      <title>{'todo'}</title>
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="initial-scale=1.0, width=device-width" />
+    </Head>
+      <div>{children}</div>
+    </div>
 }
